+++ library/src/main/java/eu/kanade/tachiyomi/network/interceptor/RateLimitInterceptor.kt
@@ -23,6 +23,7 @@ import kotlin.time.Duration.Companion.seconds
  * @param period [Long]   The limiting duration. Defaults to 1.
  * @param unit [TimeUnit] The unit of time for the period. Defaults to seconds.
  */
+@Deprecated("Use the version with kotlin.time APIs instead.")
 fun OkHttpClient.Builder.rateLimit(
     permits: Int,
     period: Long = 1,
@@ -45,49 +46,4 @@ fun OkHttpClient.Builder.rateLimit(
 fun OkHttpClient.Builder.rateLimit(
     permits: Int,
     period: Duration = 1.seconds
-): OkHttpClient.Builder = throw Exception("Stub!")
-
-/**
- * An OkHttp interceptor that handles given url host's rate limiting.
- *
- * This uses Java Time APIs and is the legacy method, kept
- * for compatibility reasons with existing extensions.
- *
- * Examples:
- *
- * httpUrl = "api.manga.com".toHttpUrlOrNull(), permits = 5, period = 1, unit = seconds  =>  5 requests per second to api.manga.com
- * httpUrl = "imagecdn.manga.com".toHttpUrlOrNull(), permits = 10, period = 2, unit = minutes  =>  10 requests per 2 minutes to imagecdn.manga.com
- *
- * @since extension-lib 1.3
- *
- * @param httpUrl [HttpUrl] The url host that this interceptor should handle. Will get url's host by using HttpUrl.host()
- * @param permits [Int]     Number of requests allowed within a period of units.
- * @param period [Long]     The limiting duration. Defaults to 1.
- * @param unit [TimeUnit]   The unit of time for the period. Defaults to seconds.
- */
-fun OkHttpClient.Builder.rateLimitHost(
-    httpUrl: HttpUrl,
-    permits: Int,
-    period: Long = 1,
-    unit: TimeUnit = TimeUnit.SECONDS,
-): OkHttpClient.Builder = throw Exception("Stub!")
-
-/**
- * An OkHttp interceptor that handles given url host's rate limiting.
- *
- * Examples:
- *
- * httpUrl = "api.manga.com".toHttpUrlOrNull(), permits = 5, period = 1.seconds =>  5 requests per second to api.manga.com
- * httpUrl = "imagecdn.manga.com".toHttpUrlOrNull(), permits = 10, period = 2.minutes  =>  10 requests per 2 minutes to imagecdn.manga.com
- *
- * @since extension-lib 1.5
- *
- * @param httpUrl [HttpUrl] The url host that this interceptor should handle. Will get url's host by using HttpUrl.host()
- * @param permits [Int]     Number of requests allowed within a period of units.
- * @param period [Duration] The limiting duration. Defaults to 1.seconds.
- */
-fun OkHttpClient.Builder.rateLimitHost(
-    httpUrl: HttpUrl,
-    permits: Int,
-    period: Duration = 1.seconds,
 ): OkHttpClient.Builder = throw Exception("Stub!")
